fix(eslint): fix elsint for deployment to vercel
