+++ postcss.config.mjs
@@ -1,6 +1,8 @@
-export default {
+const postcssConfig = {
   plugins: {
     tailwindcss: {},
     autoprefixer: {},
   },
-}
+};
+
+export default postcssConfig;
