+++ src/app/page.tsx
@@ -1,13 +1,13 @@
 import React from "react";
-import Header from "./components/Header";
-import HeroSection from "./components/HeroSection";
-import FeaturesSection from "./components/FeaturesSection";
-import MarqueSection from "./components/MarqueSection";
-import ProductSection from "./components/ProductSection";
-import TestimonialsSection from "./components/TestimonialSection";
-import BlogSection from "./components/BlogSection";
-import CTASection from "./components/CTA";
-import Footer from "./components/Footer";
+import Header from "../components/Header";
+import HeroSection from "../components/HeroSection";
+import FeaturesSection from "../components/FeaturesSection";
+import MarqueSection from "../components/MarqueSection";
+import ProductSection from "../components/ProductSection";
+import TestimonialsSection from "../components/TestimonialSection";
+import BlogSection from "../components/BlogSection";
+import CTASection from "../components/CTA";
+import Footer from "../components/Footer";
 
 const LandingPage: React.FC = () => {
   return (
